fixup! ASoC: SOF: Intel: start SoundWire links earlier for LNL+ devices
In all existing platforms, the SoundWire interrupts are re-enabled in
the post_fw_run callback.

For LNL+, we also need to re-enable this step, but since there are no
dependencies on firmware we can re-enable the interrupts in the resume
stage.

Note that we could have unconditionally re-enabled the SoundWire
interrupts in the resume stage, but to avoid breaking older platforms
with well-intended patches we do so ONLY for LNL+ devices.

Closes: https://github.com/thesofproject/linux/issues/4728
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>